Event based routing POC
